requirements.txt
biopython>=1.81
matplotlib>=3.7
numpy>=1.24
pandas>=2.0
requests>=2.28
scipy>=1.11
seaborn>=0.12
tqdm>=4.65
